componetize the article comments
* add the data endpoint for the article scoped comments
* refactored the article-scoped comments a bit so we can reuse the comment card
* comment card has been moved to components (moving towards reusability)

The article scoped comments presented a challenge because we had to somehow send it information that it can use to get the correct url. For rails it's easy, because we already know from the server side during the generation of the html.

However, since at this point the client is separated from the server context, there needs to be a way to pass that information. In this case, I wrapped the article into a div and used a data attribute on that div to give identity to the article.

This way, during the time the vue app is mounted, it can look at the dom and figure out which article context it's working with.