create prototype for the comment box
The biggest issue here is the handling of rails csrf. We need to get the csrf meta tags from the document rendered by the server and send that in along with the form payload.

We reuse the same technique from the latest comments component to send in the article id as well as the csrf token as information  into the application.